Se modificaron cosas que estuvimos haciendo anoche con manu, ya que al producirse overflow luego de una inserción, para hacer redistribuciones no se necitan saber los bytes requeridos por un nodo, sino que se necesita conocer los bytes sobrantes del nodo en el que hubo overflow.

+++ CapaIndices/BStar/BStarTree.h
@@ -7,7 +7,7 @@ class BStarTree {
 	
 	public:
 		//Constructor que recibe el tamaño de los nodos (en bytes)
-		BStarTree(unsigned tamanioNodo);
+		BStarTree(unsigned short tamanioNodo);
 		
 		virtual ~BStarTree();
 		
@@ -41,6 +41,7 @@ class BStarTree {
 									  NodoBStar* &nodoHnoDer) const;
 		
 		//Pasa una clave de un hermano derecho a otro nodo
+		//Se usa en la eliminación de una clave
 		void pasarClaveDerecha(NodoBStar* nodoDestino, NodoBStar* nodoPadre,
 							   NodoBStar* nodoHnoDer);
 		
@@ -48,6 +49,7 @@ class BStarTree {
 											  NodoBStar* &nodoHnoIzq) const;
 		
 		//Pasa una clave de un hermano izquierdo a otro nodo
+		//Se usa en la eliminación de una clave
 		void pasarClaveIzquierda(NodoBStar* nodoDestino, NodoBStar* nodoPadre,
 				   NodoBStar* nodoHnoIzq);
